requirements.txt
Flask
tensorflow
pillow
numpy
protobuf
Gunicorn
